Rework file system in server, add dir for tests
