+++ httpserver/httpserver_src/.gitignore
@@ -1,3 +1,2 @@
-.vs/
 bin/
 obj/
