--- server/HTTP-Server.csproj
@@ -1,8 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
-    <RootNamespace>HTTP_Server</RootNamespace>
-  </PropertyGroup>
-
-</Project>
