--- server/HTTP-Server.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.30804.86
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HTTP-Server", "HTTP-Server.csproj", "{5C043D9E-00E2-4591-BD3A-77B203134C17}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{5C043D9E-00E2-4591-BD3A-77B203134C17}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{5C043D9E-00E2-4591-BD3A-77B203134C17}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{5C043D9E-00E2-4591-BD3A-77B203134C17}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{5C043D9E-00E2-4591-BD3A-77B203134C17}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {E8DE2662-4ADF-4CBC-A901-C01DE1E82820}
-	EndGlobalSection
-EndGlobal
